Move users read handler to separate file
